+++ src/app/locales/en/components/filesystem-browser/table-row-columns.js
@@ -1,0 +1,4 @@
+export default {
+  copy: 'copying',
+  move: 'moving',
+};