+++ src/app/locales/en/utils/archive-browser-model.js
@@ -1,12 +1,15 @@
 import BaseBrowserModel from './base-browser-model';
 import _ from 'lodash';
 
+export const createIncrementalArchive = 'Create incremental archive';
+
 export default _.merge({}, BaseBrowserModel, {
   startingDownload: 'starting archive download',
   archiveList: 'Archives list',
   alreadyPurging: 'Not available for archives that are being purged.',
   fileActions: {
     createArchive: 'Create archive',
+    createIncrementalArchive,
     downloadTar: 'Download (tar)',
     copyArchiveId: 'Copy archive ID',
     purge: {