+++ CHANGELOG.md
@@ -6,6 +6,10 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-7371 Added showing progress for copied or moved operation
+* VFS-8031 Added select current directory button in items selector (if applicable)
+* VFS-8212 Fixed not working current directory rename
+* VFS-7930 Added "create incremental archive" in archive context menu
 * VFS-7896 Added workflow and task execution audit log preview
 * VFS-8102 Optimized notifying about upload progress to Onezone GUI
 * VFS-8086 Jumping to workflow execution details after execution start